ADD DISPATCH BEFORE DISPLAY PARTY:
it works.
