+++ front/app/src/Views/MyActiveGame/MyActiveGame.tsx
@@ -7,20 +7,12 @@ import {
 	Stack,
 	Button,
 	Box,
-	Card,
-	CardActions,
-	CardContent,
-	CardMedia,
-	Grid
 } from "@mui/material";
 import MenuBar from "../../Component/MenuBar/MenuBar";
-import { useAppDispatch, useAppSelector } from "../../Redux/hooks/redux-hooks";
+import { useAppDispatch } from "../../Redux/hooks/redux-hooks";
 import {
 	getMyActiveGame,
-	revokeGameWithUuid
 } from "../../Redux/store/gameEngineAction";
-import ButtonRemoveGame from "./ButtonRemoveGame";
-import GamePreview from "./GamePreview";
 
 const	Header = () =>
 {
@@ -86,320 +78,6 @@ const	Header = () =>
 const cardsAlone = [];
 
 
-type ButtonJoinPartyProps = {
-	playerOne: any;
-	playerTwo: any;
-};
-const	ButtonJoinParty = (props: ButtonJoinPartyProps) =>
-{
-	const	myProfileId = useAppSelector((state) =>
-	{
-		return (state.controller.user.id);
-	});
-	let	amIConnected;
-	let message;
-	amIConnected = false;
-	if (props.playerOne.profileId === myProfileId)
-	{
-		if (props.playerOne.socketId === "undefined")
-		{
-			message = "Commencer la partie";
-		}
-		else if (props.playerOne.socketId === "invited")
-		{
-			message = "Commencer la partie";
-		}
-		else if (props.playerOne.socketId === "disconnected")
-		{
-			message = "Votre adversaire vous attend";
-		}
-		else if (props.playerOne.socketId === "revoked")
-		{
-			message = "Fin de partie";
-		}
-		else
-		{
-			message = "Vous etes deja en partie";
-			amIConnected = true;
-		}
-	}
-	if (props.playerTwo.profileId === myProfileId)
-	{
-		if (props.playerTwo.socketId === "undefined") // must not be on friends
-		{
-			message = "Commencer la partie";
-		}
-		else if (props.playerTwo.socketId === "invited")
-		{
-			message = "Commencer la partie";
-		}
-		else if (props.playerTwo.socketId === "disconnected")
-		{
-			message = "Votre adversaire vous attend";
-		}
-		else if (props.playerTwo.socketId === "revoked")
-		{
-			message = "Fin de partie";
-		}
-		else
-		{
-			message = "Vous y jouez deja";
-			amIConnected = true;
-		}
-	}
-	let	render;
-	if (amIConnected === true)
-		render = <Button size="small" disabled={true}>{message}</Button>;
-	else
-		render = <Button size="small">Entrer en jeu</Button>;
-	return (render);
-};
-
-
-const	FriendsParty = () =>
-{
-	const	arrayGameFriends = useAppSelector((state) =>
-	{
-		return (state.gameEngine.myGameActive.friend);
-	});
-
-	return (
-		<Container sx={{ py: 8 }} maxWidth="md">
-			<Typography
-				component="h2"
-				variant="h3"
-				align="center"
-				color="text.primary"
-				gutterBottom
-			>
-				Parties entre amis
-			</Typography>
-			<Grid container spacing={4}>
-			{
-				arrayGameFriends.map((game: any, index: number) =>
-				{
-					const	userConnected = game.userConnected;
-					const	gameMode = game.gameMode;
-					const	uuid = game.uuid;
-					const	roomName = game.roomName;
-					const	playerOne = game.playerOne;
-					const	playerTwo = game.playerTwo;
-					const	board = game.board;
-					const 	loop = game.loop;
-					const	ball = game.ball;
-					const	revoked = game.revoked;
-
-					const	score = game.playerOne.score
-						+ ":" + game.playerTwo.score;
-					// check doublon dans les loop et GameServe
-					let elapsedTime;
-					if (loop.frameNumber === 0)
-						elapsedTime = "aucun";
-					else
-					{
-						elapsedTime
-							= loop.frameNumber / loop.frameRate + " sec.";
-					}
-					let userOneAvatar;
-					if (playerOne.profileId !== "undefined")
-						userOneAvatar = playerOne.profilePicture;
-					else
-						userOneAvatar = 'http://localhost:3000/cdn/image/profile/default.png';
-					let userTwoAvatar;
-						if (playerTwo.profileId !== "undefined")
-							userTwoAvatar = playerTwo.profilePicture;
-						else
-							userTwoAvatar = 'http://localhost:3000/cdn/image/profile/default.png';
-					console.log("please use correct call to URL :) ",
-						userOneAvatar + " " + userTwoAvatar);
-					console.log(game);
-					return (
-						<Grid item key={index} xs={12} sm={6} md={4}>
-							<Card
-								sx={{
-									height: "100%",
-									display: "flex",
-									flexDirection: "column"
-								}}
-							>
-								<Grid
-									container
-									sx={{
-										flexDirection: "row",
-										border: "1px solid blue"
-									}}
-								>
-									{/* // player one  */}
-									<Grid item xs={6}
-										sx={{
-											border: "1px solid blue"
-										}}
-									>
-										<CardMedia
-											component="div"
-											sx={{
-												// 16:9
-												// pt: "56.25%",
-												pt: "100%",
-											}}
-											image={userOneAvatar}
-										/>
-									</ Grid>
-									{/* // player Two */}
-									<Grid item xs={6}
-										sx={{
-											border: "1px solid blue"
-										}}
-									>
-										<CardMedia
-											component="div"
-											sx={{
-												// 16:9
-												// pt: "56.25%",
-												pt: "100%",
-											}}
-											image={userTwoAvatar}
-										/>
-									</ Grid>
-								</ Grid>
-								<GamePreview
-									ball={ball}
-									board={board}
-									playerOne={playerOne}
-									playerTwo={playerTwo}
-								/>
-								<CardContent sx={{ flexGrow: 1 }}>
-									<Typography
-										gutterBottom
-										variant="h5"
-										component="h2"
-									>
-										Instance: {roomName} <br />
-									</Typography>
-									<Typography>
-										Partie en mode: "{gameMode}"
-										<br />
-										Score : {score}
-										<br />
-										Temps passe en jeu : {elapsedTime}
-										<br />
-										Joueur dans le salon : {userConnected}
-									</Typography>
-								</CardContent>
-								<CardActions>
-									<ButtonJoinParty
-										playerOne={playerOne}
-										playerTwo={playerTwo}
-									/>
-									<ButtonRemoveGame
-										uuid={uuid}
-										revoked={revoked}
-									/>
-								</CardActions>
-							</Card>
-						</Grid>
-					);
-				})
-			}
-			</Grid>
-		</Container>
-	);
-};
-
-const	PrivateParty = () =>
-{
-	return (
-		<Container sx={{ py: 8 }} maxWidth="md">
-			<Typography
-				component="h2"
-				variant="h3"
-				align="center"
-				color="text.primary"
-				gutterBottom
-			>
-				Partie random
-			</Typography>
-			<Grid container spacing={4}>
-			{
-				cardsAlone.map((card) =>
-				{
-					return (
-						<Grid item key={card} xs={12} sm={6} md={4}>
-							<Card
-								sx={{
-									height: "100%",
-									display: "flex",
-									flexDirection: "column"
-								}}
-							>
-								<Grid
-									container
-									sx={{
-										flexDirection: "row",
-										border: "1px solid blue"
-									}}
-								>
-									{/* // player one  */}
-									<Grid item xs={6}
-										sx={{
-											border: "1px solid blue"
-										}}
-									>
-										<CardMedia
-											component="div"
-											sx={{
-												// 16:9
-												// pt: "56.25%",
-												pt: "100%",
-											}}
-											image="https://source.unsplash.com/random?wallpapers"
-										/>
-									</ Grid>
-									{/* // player Two */}
-									<Grid item xs={6}
-										sx={{
-											border: "1px solid blue"
-										}}
-									>
-										<CardMedia
-											component="div"
-											sx={{
-												// 16:9
-												// pt: "56.25%",
-												pt: "100%",
-											}}
-											image="https://source.unsplash.com/random?wallpapers"
-										/>
-									</ Grid>
-								</ Grid>
-								<GamePreview />
-								<CardContent sx={{ flexGrow: 1 }}>
-									<Typography
-										gutterBottom
-										variant="h5"
-										component="h2"
-									>
-										Heading
-									</Typography>
-									<Typography>
-										This is a media card.
-										You can use this section to describe the
-										content.
-									</Typography>
-								</CardContent>
-								<CardActions>
-									<Button size="small">Entrer en jeu</Button>
-									<Button size="small">Supprimer</Button>
-								</CardActions>
-							</Card>
-						</Grid>
-					);
-				})
-			}
-			</Grid>
-		</Container>
-	);
-};
 
 const	MyActiveGame = () =>
 {
@@ -407,8 +85,8 @@ const	MyActiveGame = () =>
 		<>
 			<MenuBar />
 			<Header />
-			<FriendsParty />
-			<PrivateParty />
+			{/* <FriendsParty /> */}
+			{/* <PrivateParty /> */}
 		</>
 	);
 };
